web/requirements.txt
azure-common==1.1.28
azure-core==1.26.4
azure-mgmt-compute==29.1.0
azure-mgmt-core==1.4.0
azure-mgmt-monitor==6.0.0
azure-nspkg==3.0.2
azure-servicemanagement-legacy==0.20.7
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
boto3==1.26.123
botocore==1.29.123
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
cssselect==1.2.0
cssutils==2.6.0
dnspython @ git+https://github.com/rthalley/dnspython@925a669f0f15f00908d3e066dbbc461852c06819
Flask==2.2.5
Flask-Cors==3.0.10
Flask-Login==0.6.2
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud-compute==1.11.0
google-cloud-container==2.21.0
google-cloud-core==2.3.2
google-cloud-language==2.9.1
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.54.0
grpcio-status==1.54.0
httplib2==0.22.0
hurry==1.1
hurry.filesize==0.9
idna==3.4
importlib-metadata==6.6.0
install==1.3.5
isodate==0.6.1
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.1.2
jmespath==1.0.1
keyring==23.13.1
kubernetes==26.1.0
language==0.6
lxml==4.9.2
MarkupSafe==2.1.2
more-itertools==9.1.0
msrest==0.7.1
multi-key-dict==2.0.3
oauth2client==4.1.3
oauthlib==3.2.2
paramiko==3.1.0
pbr==5.11.1
Pillow==9.5.0
premailer==3.10.0
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.6.0
pymongo==4.3.3
PyNaCl==1.5.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jenkins==1.8.0
PyYAML==6.0
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
soupsieve==2.4.1
storage==0.0.4.3
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.25.4
websocket-client==1.5.1
Werkzeug==2.3.2
yagmail==0.15.293
zipp==3.15.0
